HBX-2390: Create a JBoss Tools adaptation layer in Hibernate Tools   - Move test class 'HbmExporterExtTest' and implementation class 'HbmExporterExt' from 'org.hibernate.tool.orm.jbt.util' to 'org.hibernate.tool.orm.jbt.wrp'
